fix double-click for history box text selection
The history box footer wasn't showing up when selecting the words
with a double-click.

It's because on the first click, an animation is launched to hide
it (just a few lines below). After this 0.3s animation, there's a
call to .hide(). Then, if the double-click takes less than 0.3s,
the .hide() call is called after the .show() call.

minor: fixed a typo on the top-level comment also